factor mesh publication presets
- Add a mesh publication preset helper for host-control files
- Route GitHub Pages .nojekyll creation through the publication profile
- Remove the legacy includeNoJekyll mesh-create bypass
- Update mesh-create tests for host-neutral core bootstrap
