+++ src/core/mesh/mod.ts
@@ -1,1 +1,2 @@
 export * from "./create.ts";
+export * from "./publication_preset.ts";
